+++ cocos/editor-support/cocostudio/CCDataReaderHelper.cpp
@@ -291,7 +291,7 @@ void DataReaderHelper::addDataFromFile(const char *filePath)
     size_t startPos = filePathStr.find_last_of(".");
     std::string str = &filePathStr[startPos];
 
-    long size;
+    ssize_t size;
     std::string fullPath = CCFileUtils::getInstance()->fullPathForFilename(filePath);
     char *pFileContent = (char *)CCFileUtils::getInstance()->getFileData(fullPath.c_str() , "r", &size);
 
@@ -391,7 +391,7 @@ void DataReaderHelper::addDataFromFileAsync(const char *imagePath, const char *p
     std::string str = &filePathStr[startPos];
 
     std::string fullPath = CCFileUtils::getInstance()->fullPathForFilename(filePath);
-    long size;
+    ssize_t size;
 
     // XXX fileContent is being leaked
     data->fileContent = (char *)CCFileUtils::getInstance()->getFileData(fullPath.c_str() , "r", &size);
